fix: support Mod Organizer 2.
When launching with MO2, std::weakly_canonical is not behaving correctly. Security layer is now denying access when / or \ is present in child path, only with MO2.
Only use path from std::directory_iterator entries when listing files. Building a custom path was getting in the way when using MO2.